% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplicon_extract.R
\name{find_amplicon}
\alias{find_amplicon}
\title{Find the location of an amplicon in an alignment}
\usage{
find_amplicon(
  aln,
  fwd_primer,
  rev_primer,
  trim = c("none", "retain", "remove"),
  mark = TRUE,
  fwd_char = "{",
  rev_char = "}",
  outfile = NULL
)
}
\arguments{
\item{aln}{(\code{connection}, \code{character} string giving a file name in Stockholm format,
\code{\link[Biostrings:MultipleAlignment-class]{DNAMultipleAlignment}},
\code{\link[Biostrings:MultipleAlignment-class]{RNAMultipleAlignment}},
\code{\link[Biostrings:XStringSet-class]{DNAStringSet}},
\code{\link[Biostrings:XStringSet-class]{RNAStringSet}},
\code{\link[inferrnal:StockholmMultipleAlignment-class]{StockholmMultipleAlignment}},
or \code{character} vector)
DNA or RNA multiple alignment in which to search for an amplicon.}

\item{fwd_primer}{(\code{character} string or \code{\link[Biostrings:XString-class]{DNAString}})
Forward primer sequence to define the target amplicon.}

\item{rev_primer}{(\code{character} string or \code{\link[Biostrings:XString-class]{DNAString}})
Reverse primer sequence to define the target amplicon. Should be given from
5' to 3' in the primer; i.e. the reverse complement of the expected sequence
in the alignment.}

\item{trim}{(one of \code{"none"}, \code{"retain"}, or \code{"remove"})
Choice of how to trim the alignment: if \code{"none"} then the alignment is not
trimmed; if \code{"retain"} then the alignment is trimmed to the amplicon,
including the primer sites; if \code{"remove"} then the alignment is trimmed to
the amplicon and the primer sites are also removed.}

\item{mark}{(\code{logical} flag)
If \code{TRUE} (default) the primer sites are marked in the alignment RF line.}

\item{fwd_char}{(single \code{character})
Character to use for marking the forward primer location in the RF line.}

\item{rev_char}{(single \code{character})
Character to use for marking the reverse primer location in the RF line.}

\item{outfile}{(\code{character} file name or \code{\link{connection}})
If non-\code{NULL}, an output file or connection to write the result to.}
}
\value{
\code{\link[inferrnal:StockholmMultipleAlignment-class]{StockholmMultipleAlignment}}
object with modified RF line to mark the primer locations, or if \code{outfile} is
given, \code{NULL} invisibly.
}
\description{
Find the location of an amplicon in an alignment
}
