% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplicon_extract.R
\name{gap_fill}
\alias{gap_fill}
\title{Convert range from ungapped sequence to gapped sequence}
\usage{
gap_fill(rstart, rend, gaps)
}
\arguments{
\item{rstart}{(\code{integer} scalar)
the start point of the range}

\item{rend}{(\code{integer} scalar)
the end point of the range}

\item{gaps}{(\code{\link[IRanges:IRanges-constructor]{IRanges}} object)
the locations of gaps in the gapped sequence.}
}
\value{
named \code{integer} of length two, with elements \code{"start"} and \code{"end"},
giving the coordinates in the gapped sequence which correspond to coordinates
\code{"rstart"} and \code{"rend"} in the ungapped sequence.
}
\description{
Convert range from ungapped sequence to gapped sequence
}
