% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsux.R
\docType{data}
\name{itsx_result}
\alias{itsx_result}
\title{Sample ITSx results from difficult sequences.}
\format{
An object of class \code{list} of length 6.
}
\usage{
itsx_result
}
\description{
The sequences are a selection of the sample sequences from \code{inferrnal},
specifically elements \code{[c(1,4,20,32,33,43,46,49)]}.  These were chosen
particularly to give bad 5.8S detection results, and are not representative
of normal results from ITSx.
}
\keyword{datasets}
