% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsux.R
\name{lsux}
\alias{lsux}
\title{Extract variable/conserved domains from LSU rDNA}
\usage{
lsux(
  seq,
  cm_5.8S = system.file(file.path("extdata", "RF00002.cm"), package = "inferrnal"),
  cm_32S = system.file(file.path("extdata", "fungi_32S.cm"), package = "LSUx"),
  glocal = TRUE,
  ITS1 = FALSE,
  cpu
)
}
\arguments{
\item{seq}{(filename readable by Infernal,
\code{\link[Biostrings]{XStringSet}},
\code{\link[ShortRead]{ShortRead}},
or \code{character} vector) sequences
to extract regions from}

\item{cm_5.8S}{(filename) covariance model for 5.8S rRNA}

\item{cm_32S}{(filename) covariance model for 32S pre-rRNA
(5.8S, ITS2, and LSU)}

\item{glocal}{(\code{logical} scalar) if \code{TRUE}, use glocal alignment in
\code{\link[inferrnal]{cmalign}}}

\item{ITS1}{(\code{logical} scalar) if \code{TRUE}, include sequence fragment
before 5.8S (if any) as ITS1}

\item{cpu}{(\code{integer} scalar) number of threads to use in Infernal calls}
}
\value{
a \code{\link[tibble]{tibble}} with one row for each region found for
 each input sequence.
 The columns are: \describe{
 \item{\code{seq_id} (\code{character})}{ the sequence name from
     \code{seq}}
 \item{\code{length} (\code{integer})}{ the length of the original sequence
      in base pairs}
 \item{\code{region} (\code{character})}{the name of the found domain}
 \item{\code{*_start} (\code{integer})}{the starting base for that domain in
   this sequence. Can be \code{"5_8S"}, \code{"ITS2"}, \code{"LSU1"},
   \code{"V2"}, \code{"LSU2"}, \code{"V3"}, etc.}
 \item{\code{end} (\code{integer})}{as \code{start}, but giving the end
     base for the domain.}}
}
\description{
Extracts alternating variable and conserved domains from the contiguous rDNA
regions which form the eukaryotic ribosomal large subunit,  i.e. 5.8S RNA,
ITS2, and 28S RNA. For the purposes of this document, this region will be
referred to as the 32S precursor RNA, as in humans, although its actual size
in Svedberg units varies between lineages.
}
\details{
Input sequences should contain, at a minimum, a significant fraction of the
5.8S RNA, which is used to define the 5' end of 32S. Any base pairs before
the 5' end of 5.8S will be considered to be ITS1 (\code{ITS1 = TRUE}) or
discarded (\code{ITS1 = FALSE}).  Input sequences should not extend past the
end of the 32S model at the 3' end.

LSUx requires two covariance models: one for 5.8S, which is used in
\code{\link[inferrnal]{cmsearch}}, and one for 32S, which is used in
\code{\link[inferrnal]{cmalign}}.

The 5.8S model can be
\href{https://rfam.xfam.org/family/RF00002}{RF00002 from Rfam} (the default),
or an equivalent.  It must be calibrated using \code{cmcalibrate} from
Infernal.

The 32S model must include annotations in the reference line ("#=GC RF" in
the seed alignment) to distinguish conserved and variable regions. The
annotations should be sequential characters in the range \code{"1..9A..Z"}
for conserved domains, \code{"v"} for variable domains, and \code{"."} for
unaligned gaps in the seed alignment.
In the output, the conserved domains will be named "5_8S", "LSU1", "LSU2",
...; the variable domains will be named "ITS2", "V1", "V2", ...

Two example models are included,
both based on the
\href{https://github.com/rdpstaff/fungene_pipeline/blob/master/resources/RRNA_28S/model.cm}{RDP fungal LSU CM},
and annotated with variable regions according to Raué (1988).
The first, \code{system.file(file.path("extdata", "fungal_32S.cm"),
package = "LSUx")}, includes the full LSU region.  The second,
\code{system.file(file.path("extdata", "fungal_32S_LR5.cm"),
package = "LSUx")}, is truncated at the binding site of the LR5 primer, and
should be faster for input sequences which do not extend past that point.
The seed alignments are also provided.

If generating similar truncated alignments with different endpoints, it is
critical to remove unpaired secondary structure elements from the
\code{"#=GC SS_cons"} line of the seed alignment.
}
\examples{
# the sample data was amplified with primers ITS1 and LR5, so the truncated
# cm is appropriate.
seq <- system.file("extdata/sample.fasta", package = "inferrnal")
cm_32S_trunc <- system.file(
    file.path("extdata", "fungi_32S_LR5.cm"),
    package = "LSUx"
)
lsux(seq, cm_32S = cm_32S_trunc, ITS1 = TRUE, cpu = 1)
}
