% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsux.R
\name{repair_unmatched_secondary_structure}
\alias{repair_unmatched_secondary_structure}
\title{Replaced unmatched brackets in a dot-bracket RNA secondary structure}
\usage{
repair_unmatched_secondary_structure(ss)
}
\arguments{
\item{ss}{(\code{character} scalar) the secondary structure to repair.}
}
\value{
(\code{character} scalar) the repaired secondary structure.
}
\description{
Replaced unmatched brackets in a dot-bracket RNA secondary structure
}
\examples{
    # some brackets are unmatched due to truncation
    ss <- "..(((.....[[...<<<<<___>>>>>.[[.<<<___>>>.]"
    repair_unmatched_secondary_structure(ss)
}
