% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsux.R
\name{truncate_alignment}
\alias{truncate_alignment}
\title{Truncate an annotated Stockholm alignment file with secondary structure}
\usage{
truncate_alignment(aln, outfile = NULL, start = 1L, stop = 1000000L)
}
\arguments{
\item{aln}{(\code{character} filename, \code{\link{connection}}, or
\code{\link[inferrnal:StockholmMultipleAlignment-class]{StockholmMultipleAlignment}})
the stockholm alignment to be truncated.}

\item{outfile}{(\code{character} filename or \code{\link{connection}})
path or connection to output the truncated alignment to. The default, \code{NULL},
instead returns the truncated alignment as an R object.}

\item{start}{(\code{integer} scalar)
first base to include in the truncated alignment.}

\item{stop}{(\code{integer} scalar)
last base to include in the truncated alignment.}
}
\value{
the truncated alignment as a
\code{\link[inferrnal:StockholmMultipleAlignment-class]{StockholmMultipleAlignment}}
object, or if outfile is given then \code{NULL} (invisibly).
}
\description{
Simply truncating a Stockholm alignment file that includes secondary
structure annotations may result in invalid base pairing, when only one half
of a pair is removed. This function converts these half-pairs to "X",
indicating unpairable bases.
}
\examples{
aln <- system.file(file.path("extdata", "fungi_32S.stk"), package = "LSUx")
truncate_alignment(aln, tempfile("trunc", fileext = ".stk"), 1, 500)
}
